Feat (ThemeToggle) + Project organization
- Created components folder to apply single responsibility and separate of concerns.
- Implemented toggle switch feature
- Applied pre-existing main.js logic and added to ThemeToggle.jsx.
- Removed duplicate background color and fixed unnecessary gap that shows during theme transitioning.
- Renamed css background color variable name.
