+++ client/src/components/MovieCard.jsx
@@ -1,4 +1,4 @@
-export const MovieFeaturedList = ({ title, image, year, rating }) => {
+export const MovieCard = ({ title, image, year, rating }) => {
 	// Check to see if image source is avaliable. If not, set it to null
 	const imageUrl = image ? `https://image.tmdb.org/t/p/w500${image}` : null;
